refactor: 가독성을 위한 디렉토리, 메서드 네이밍 변경
