+++ src/js/index.js
@@ -1,4 +1,4 @@
-import Lotto from './lib/Lotto.js';
+import Lotto from './model/Lotto.js';
 import { $ } from './utils/querySelector.js';
 import { handlePurchaseResultToggle } from './handler/handlePurchaseResultToggle.js';
 import { handlePurchasePriceInput } from './handler/handlePurchasePriceInput.js';
